Boost name score in search queries.
Searches are almost always for var names, but
hits in the name field were underweighted,
while hits including the query in the body and
arg list scored much higher. A search for "map",
e.g., returns "dissoc" as the top hit, since it
operates on maps and uses "map" as an arg.

+ Add boost at query time: fuzzy matches in the
  name field add a score boost of 2.0.